refactor(engine): give the init deadline a home outside the wwjs adapter
The deadline SessionService races engine.initialize() against applies to every
engine, but it was derived inline from resolveAuthTimeoutMs(), which the
whatsapp-web.js adapter owned. So the session lifecycle imported one specific
engine's module to size a value it applies to all of them — and that import
also pinned the wwjs adapter lane behind this branch, since session.service.ts
is already heavily changed here.

The env parse and the derivation now sit together in engine/engine-init-timeout.ts,
where the reason the outer deadline must clear the inner auth wait can be stated
once instead of as a comment above one expression. The adapter re-exports
resolveAuthTimeoutMs, which is how callers configuring whatsapp-web.js should
still reach it.

Pure code motion: the derived deadline is unchanged, and both
whatsapp-web-js.adapter.spec.ts and session.service.spec.ts pass with no edits
(617 tests). WWEBJS_AUTH_TIMEOUT_MS still feeds the shared floor — documented
rather than changed, because it can only raise the deadline above 60s, never
lower it, so a Baileys session gets a more generous window and never a shorter
one. Separate per-engine windows would need their own env var and a behaviour
change.
